Transformer input id supports int32 & mv beam search v2 (#905)
* support int32

* mv beam search v2

+++ paddlenlp/ops/faster_transformer/transformer/faster_transformer.py
@@ -108,7 +108,7 @@ class FasterTransformer(TransformerModel):
             src_word == self.bos_id,
             dtype=paddle.get_default_dtype()).unsqueeze([1, 2]) * -1e9
         src_pos = paddle.cast(
-            src_word != self.bos_id, dtype="int64") * paddle.arange(
+            src_word != self.bos_id, dtype=src_word.dtype) * paddle.arange(
                 start=0, end=src_max_len)
 
         # Run encoder
